Move resolve into tool submodules
