+++ crates/volta-core/src/lib.rs
@@ -17,7 +17,6 @@ pub mod monitor;
 pub mod path;
 pub mod platform;
 pub mod project;
-pub(crate) mod resolve;
 pub mod run;
 pub mod session;
 pub mod shell;
